requirements.txt
streamlit
psycopg2-binary
pymongo
mysql-connector-python
pyodbc
pandas
google-generativeai
pymssql
bson
